mv solidity stuff out
add solidity file
